% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmesImp.R
\docType{data}
\name{AmesImp}
\alias{AmesImp}
\title{A Subset of the Ames Data Set with Imputed Values}
\format{A data frame with 2047 observations on 74 variables.}
\source{
\url{https://ww2.amstat.org/publications/jse/v19n3/Decock/DataDocumentation.txt}
}
\usage{
AmesImp
}
\description{
A randomly selected subset of the Ames data set.
The dataset has had its values imputed and any remaining NA values removed.
}
\examples{
str(AmesImp)
plot(AmesImp$Neighborhood, y = AmesImp$SalePrice)
}
\keyword{datasets}
