% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmesImpFctrs.R
\docType{data}
\name{AmesImpFctrs}
\alias{AmesImpFctrs}
\title{A Subset of the Ames Data Set with Imputed Values Only Including Factor Variables and Sale Price}
\format{A data frame with 2049 observations on 39 variables.}
\source{
\url{https://ww2.amstat.org/publications/jse/v19n3/Decock/DataDocumentation.txt}
}
\usage{
AmesImpFctrs
}
\description{
A randomly selected subset of the Ames data set.
The dataset contains only factor class variables and the SalePrice variable.
Missing values have been imputed.
}
\examples{
str(AmesImpFctrs)
plot(AmesImpFctrs$Neighborhood, y = AmesImpFctrs$SalePrice)
}
\keyword{datasets}
