% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmesSubset.R
\docType{data}
\name{AmesSubset}
\alias{AmesSubset}
\title{A Subset of the Ames Data Set}
\format{A data frame with 2049 observations on 82 variables.}
\source{
\url{https://ww2.amstat.org/publications/jse/v19n3/Decock/DataDocumentation.txt}
}
\usage{
AmesSubset
}
\description{
A randomly selected subset of the Ames data set.
}
\examples{
str(AmesSubset)
plot(AmesSubset$Neighborhood, y = AmesSubset$SalePrice)
}
\keyword{datasets}
