% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.bins.R
\name{tree.bins}
\alias{tree.bins}
\title{<Recategorization of Factor Variables by Decision Tree Leaves>}
\usage{
tree.bins(df = fctrs, y = SalePrice, bin.nm = "Group.", method = NULL,
  control = NULL, return = lkup.list)
}
\arguments{
\item{df}{a data.frame}

\item{y}{the response variables to be used in the rpart function}

\item{bin.nm}{the string that will be used to categorize the variables. E.g. "Group." will be used to recategorize a variable of 6 factors to 3 factors. Will result in buckets of "Group.1", "Group.2", and "Group.3"}

\item{method}{this is the method that will be used in the rpart function. If null, the default method will be used. See rpart() for further detail}

\item{control}{this is the control that will be used in the rpart function. See rpart() for further detail.}

\item{return}{this is what the function will return. There are two options: 1) lkup.list - will provide a list of the factors of each categorical variable to the new factors of each variable. 2) new.fctrs - will provide a data.frame with the recategorized categorical variables}
}
\description{
<The package takes in a data set of categorical variables and a response variable.
It creates a decision tree by using one of the categorical variables and the response variable.
The decision tree is created from the rpart function from the rpart library. The rules from the leaves of the decision tree are extracted, and used to recategorize the used variable (predictor). This step is performed for each of the categorical variables that is fed into the data component of the function. Only variables containing more than 2 factors will be considered in the function. The final output generates a data set containing the recategorized variables.>
}
\examples{
\dontrun{ "asdfsdf"}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
\keyword{factor,}
\keyword{relevel}
\keyword{rpart,}
